refactor: add utils dir.
