+++ src/main.js
@@ -1,4 +1,4 @@
-import { render } from './framework/render.js';
+import {render} from './framework/render.js';
 import DestinationsModel from './model/destinations-model.js';
 import EventsModel from './model/events-model.js';
 import OffersModel from './model/offers-model.js';
