+++ src/utils/date-utils.js
@@ -35,30 +35,7 @@ function getDateDifference(dateOne, dateTwo) {
 }
 
 
-function getCapitalizedString(string) {
-  return string[0].toUpperCase() + string.slice(1);
-}
-
-
-function getHtmlId(string) {
-  return string.replace(/\s+/g, '-').toLowerCase();
-}
-
-
-function getIdGenerator () {
-  let currentValue = 0;
-
-  return function () {
-    currentValue += 1;
-    return currentValue;
-  };
-}
-
-
 export {
   getFormattedDate,
-  getDateDifference,
-  getCapitalizedString,
-  getHtmlId,
-  getIdGenerator,
+  getDateDifference
 };
